+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.3.1.0  -- 2024-06-24
 
-* Expose SessionT, lift all functions to MonadIO
+* Introduce SessionT and MonadSession
+* Lift all functions to MonadSession
 * Add Session.getAllVersionedDocs
 
 ## 0.3.0.0  -- 2024-04-04